extract factory to request-mapper
